pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "pdm-build-skills"
description = "A pdm-backend plugin that vendors remote skill repositories into wheels."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdm-backend>=2.4",
]
authors = [
    { name = "Frost Ming", email = "me@frostming.com" },
]
license = { text = "MIT" }
keywords = ["pdm", "build", "plugin", "skills"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dynamic = ["version"]

[project.entry-points."pdm.build.hook"]
skills = "pdm_build_skills.backend:SkillsBuildHook"

[dependency-groups]
test = [
    "pytest>=8.3",
]

[tool.pdm.version]
source = "scm"

[tool.pytest.ini_options]
testpaths = ["tests"]
